[트립트립] refactor(@commons/ui): 공통 컴포넌트 리팩토링 후 불필요한 파일 삭제 및 boards-write hook 파일명 변경
