+++ maincamp/frontend/src/components/features/boards/boards-write/hook.ts
@@ -1,137 +1,112 @@
-"use client";
-
-import { useMutation } from "@apollo/client";
-import { useParams, useRouter } from "next/navigation";
-import { ChangeEvent, useState } from "react";
-import { ImageUrlArray, IUpdateBoardInput } from "./types";
-import { GraphQLError } from "graphql";
 import {
   CreateBoardDocument,
   CreateBoardInput,
   FetchBoardDocument,
+  FetchBoardQuery,
   UpdateBoardDocument,
+  UpdateBoardInput,
   UploadFileDocument,
 } from "@/commons/graphql/graphql";
-import { Modal } from "antd";
-import { Address } from "react-daum-postcode";
+import { useParams, useRouter } from "next/navigation";
+import { ChangeEvent, useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
+import { Modal } from "antd";
+import { useMutation } from "@apollo/client";
+import { Address } from "react-daum-postcode";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { GraphQLError } from "graphql";
+import {
+  boardCreateFormSchema,
+  BoardCreateFormValues,
+  boardUpdateFormSchema,
+  BoardupdateFormValues,
+} from "./schema";
 
-export default function useBoardsWrite({ data }: { data?: any }) {
+/*
+    CreateBoardInput
+
+    boardAddress?: InputMaybe<BoardAddressInput>;
+    contents: Scalars["String"]["input"];
+    images?: InputMaybe<Array<Scalars["String"]["input"]>>;
+    password?: InputMaybe<Scalars["String"]["input"]>;
+    title: Scalars["String"]["input"];
+    writer?: InputMaybe<Scalars["String"]["input"]>;
+    youtubeUrl?: InputMaybe<Scalars["String"]["input"]>;
+
+*/
+
+export default function useBoardForm({ data, isEdit }: { data?: FetchBoardQuery; isEdit?: boolean }) {
   // 0. 세팅
+  // 0-1. 라우터
   const router = useRouter();
   const params = useParams();
 
-  /* 게시물 등록 유효성 검사 */
-  // 1. 작성자, 비밀번호, 제목, 컨텐츠 작성 시 setState로 상태 변경
-  const [inputs, setInputs] = useState({
-    writer: data?.fetchBoard.writer ?? "",
-    password: data?.fetchBoard.password ?? "",
-    title: data?.fetchBoard.title ?? "",
-    contents: data?.fetchBoard.contents ?? "",
+  // 1. useForm 세팅
+  // 1-1. useForm 초기값세팅
+  const methods = useForm<BoardCreateFormValues | BoardupdateFormValues>({
+    defaultValues: {
+      writer: "",
+      password: "",
+      title: "",
+      contents: "",
+      boardAddress: {
+        zipcode: "",
+        address: "",
+        addressDetail: "",
+      },
+      youtubeUrl: "",
+      images: [undefined, undefined, undefined],
+    },
+    resolver: zodResolver(isEdit ? boardUpdateFormSchema : boardCreateFormSchema),
+    mode: "onChange",
   });
-  const [youtubeUrl, setYoutubeUrl] = useState(
-    !data ? "" : data.fetchBoard.youtubeUrl
-  );
-  const [zipcode, setZipcode] = useState(
-    !data ? "" : data.fetchBoard.boardAddress?.zipcode
-  );
-  const [address, setAddress] = useState(
-    !data ? "" : data.fetchBoard.boardAddress?.address
-  );
-  const [addressDetail, setAddressDetail] = useState(
-    !data ? "" : data.fetchBoard.boardAddress?.addressDetail
-  );
-  const [images, setImages] = useState<ImageUrlArray>(
-    !data ? [undefined, undefined, undefined] : data?.fetchBoard.images
-  );
 
-  // 1-2. 게시글 생성 API 요청 함수
-  const [createBoard] = useMutation(CreateBoardDocument);
+  const { register, handleSubmit, formState, watch, reset, getValues } = methods;
 
-  // 1-3. 게시글 수정 API 요청 함수
-  const [updateBoard] = useMutation(UpdateBoardDocument);
-
-  // 1-4. 이미지 업로드 API 요청 함수
-  const [uploadFile] = useMutation(UploadFileDocument);
-
-  // 2. 필수 작성 요소 작성 여부에 따른 버튼 활성화
-  const [isValid, setIsValid] = useState(true);
-
-  // 3. Change Event에 따른 유효성 검증
-  const onChangeInputs = (
-    event:
-      | React.ChangeEvent<HTMLInputElement>
-      | React.ChangeEvent<HTMLTextAreaElement>
-  ) => {
-    setInputs({
-      ...inputs,
-      [event.target.id]: event.target.value,
-    });
-
-    if (inputs.writer && inputs.password && inputs.title && inputs.contents) {
-      setIsValid(false);
-    } else {
-      setIsValid(true);
+  // 1-2. useForm 수정하기 페이지에서 data.fetchBoard 불러오기
+  useEffect(() => {
+    if (data?.fetchBoard) {
+      reset({
+        writer: data.fetchBoard.writer,
+        title: data.fetchBoard.title,
+        contents: data.fetchBoard.contents,
+        boardAddress: {
+          zipcode: data.fetchBoard.boardAddress?.zipcode,
+          address: data.fetchBoard.boardAddress?.address,
+          addressDetail: data.fetchBoard.boardAddress?.addressDetail,
+        },
+        images: data.fetchBoard.images,
+        youtubeUrl: data.fetchBoard.youtubeUrl,
+      });
     }
-  };
+  }, [data, reset]);
 
-  // 3-1. 필수 요소 아닌 ChangeEvent 추가
-  // (1) YoutubeUrl
-  const onChangeYoutubeUrl = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const value = event.target.value;
-    setYoutubeUrl(value);
-  };
-
-  // (2) 주소 입력 API 추가
+  // 1-3. 모달 관련 state 설정
   const [isModalOpen, setIsModalOpen] = useState(false);
 
   const onToggleModal = () => {
-    // console.log(isModalOpen)
     setIsModalOpen((prev) => !prev);
   };
 
-  const onChangeBoardAddress = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const { id, value } = event.target;
-    switch (id) {
-      case "zipcode": {
-        setZipcode(value);
-        break;
-      }
-      case "address": {
-        setAddress(value);
-        break;
-      }
-      case "addressDetail": {
-        setAddressDetail(value);
-        break;
-      }
-    }
-  };
+  // 1-4. 게시글 생성 API 요청 함수
+  const [createBoard] = useMutation(CreateBoardDocument);
 
-  const boardAddress = {
-    zipcode: zipcode,
-    address: address,
-    addressDetail: addressDetail,
-  };
+  // 1-5. 게시글 수정 API 요청 함수
+  const [updateBoard] = useMutation(UpdateBoardDocument);
 
-  const handleComplete = (data: Address) => {
-    console.log(data); // e.g. '서울 성동구 왕십리로2길 20 (성수동1가)'
-    setZipcode(data.zonecode);
-    setAddress(data.address);
-    setAddressDetail(data.buildingName);
-    onToggleModal();
-  };
+  // 1-6. 이미지 업로드 API 요청 함수
+  const [uploadFile] = useMutation(UploadFileDocument);
 
-  // (3) 파일 업로드 추가
-  const onChangeFile = async (
-    event: ChangeEvent<HTMLInputElement>,
-    index: number
-  ) => {
-    const { id, files } = event.target;
+  // 2. 함수
+  // 2-1. 파일 변경 하기
+  const onChangeFile = async (event: ChangeEvent<HTMLInputElement>, index: number) => {
+    console.log(watch("images"));
+    const files = event.target.files;
     if (!files || files.length === 0) return;
 
     const file = files?.[0];
 
+    // 파일 사이즈 유효성 검사
     if (file.size > 5 * 1024 * 1024) {
       const showErrorModal = () =>
         Modal.error({
@@ -142,6 +117,7 @@ export default function useBoardsWrite({ data }: { data?: any }) {
       return;
     }
 
+    // 파일 업로드 API
     const result = await uploadFile({
       variables: {
         file,
@@ -151,28 +127,60 @@ export default function useBoardsWrite({ data }: { data?: any }) {
     console.log(result.data?.uploadFile.url);
 
     const fileUrl = result.data?.uploadFile.url;
-    setImages((preUrls) => {
-      const newUrls = [...preUrls];
-      newUrls[index] = fileUrl;
-      return newUrls;
+
+    // 기존 값에서 변경: current(기존 createBoardInput), newImages(기존 이미지 배열 => 새로운 이미지 배열)
+    // reset시, current로 얕은 복사를 안하면 다른 내용들이 초기화..!
+    const current = getValues();
+    const newImages = watch("images") ?? [];
+
+    newImages[index] = fileUrl ?? "";
+    reset({
+      ...current,
+      images: newImages,
     });
   };
 
-  // (4) 파일 삭제 추가
+  // 2-2. 파일 삭제 하기
   const onClickDelete = (index: number) => {
-    setImages((preUrls) => {
-      const newUrls = [...preUrls];
-      newUrls[index] = undefined;
-      return newUrls;
+    const current = getValues();
+    const newImages = watch("images") ?? [];
+
+    newImages[index] = "";
+    reset({
+      ...current,
+      images: newImages,
     });
   };
 
-  // 4. 등록하기 버튼
+  // 2-3. 취소하기
+  const onClickCancel = () => {
+    router.back();
+  };
 
-  const onClickSubmit = async (data: CreateBoardInput) => {
+  // 2-4. 주소 입력하기
+  const handleComplete = (data: Address) => {
+    const current = getValues();
+    console.log(data); // e.g. '서울 성동구 왕십리로2길 20 (성수동1가)'
+    reset({
+      ...current,
+      boardAddress: {
+        zipcode: data.zonecode,
+        address: data.address,
+        addressDetail: data.buildingName,
+      },
+    });
+    onToggleModal();
+  };
+
+  // 3. 등록, 수정하기 함수
+  // 3-1. 등록하기
+  const onClickSubmit = async (data: BoardCreateFormValues) => {
     console.log(data);
-    const createBoardInput = data;
-    createBoardInput.images = data.images?.filter(Boolean);
+    const createBoardInput: CreateBoardInput = {
+      ...data,
+      images: data.images?.filter(Boolean) as string[],
+      boardAddress: data.boardAddress,
+    };
 
     console.log(createBoardInput);
     try {
@@ -189,125 +197,65 @@ export default function useBoardsWrite({ data }: { data?: any }) {
       showErrorModal();
     }
   };
-  // const onClickSubmit = async (data: CreateBoardInput) => {
-  //   console.log(data);
-  //   const newImages = data.images?.filter(Boolean)
 
-  // try {
-  //   const result = await createBoard({
-  //     variables: {
-  //       createBoardInput: data,
-  //       // createBoardInput: {
-  //       //   writer: inputs.writer,
-  //       //   password: inputs.password,
-  //       //   title: inputs.title,
-  //       //   contents: inputs.contents,
-  //       //   youtubeUrl: youtubeUrl,
-  //       //   boardAddress: {
-  //       //     zipcode: zipcode,
-  //       //     address: address,
-  //       //     addressDetail: addressDetail,
-  //       //   },
-  //       //   images: images.filter(Boolean) as string[],
-  //       // },
-  //     },
-  //   });
-  //   console.log("🚀 ~ onClickBtn ~ result:", result);
-  //   const boardId = result.data?.createBoard._id;
-  //   router.push(`/boards/${boardId}`);
-  // } catch (error) {
-  //   const showErrorModal = () =>
-  //     Modal.error({
-  //       title: "에러가 발생하였습니다.",
-  //       content: (error as string) ?? "에러가 발생하였습니다",
-  //     });
-  //   showErrorModal();
-  // }
-
-  // 5. 수정하기 버튼
-  // 수정 버튼 클릭 시 updateBoard 진행
-  const onClickUpdate = async () => {
-    // 5-1. 수정된 사항만 업데이트 될 수 있도록 variables 설정
-    const updateBoardInput: IUpdateBoardInput = {};
-    if (inputs.title !== data.fetchBoard.title && inputs.title.length > 0)
-      updateBoardInput.title = inputs.title;
+  // 3-2. 수정하기
+  const onClickUpdate = async (formData: BoardupdateFormValues) => {
+    const updateBoardInput: UpdateBoardInput = {};
+    const boardId = params.boardId as string;
+    // const values = getValues();
+    if (formData.title !== data?.fetchBoard.title) updateBoardInput.title = formData.title;
+    if (formData.contents !== data?.fetchBoard.contents) updateBoardInput.contents = formData.contents;
+    if (formData.youtubeUrl !== data?.fetchBoard.youtubeUrl)
+      updateBoardInput.youtubeUrl = formData.youtubeUrl;
     if (
-      inputs.contents !== data.fetchBoard.contents &&
-      inputs.contents.length > 0
-    )
-      updateBoardInput.contents = inputs.contents;
-    if (youtubeUrl !== data.fetchBoard.youtubeUrl)
-      updateBoardInput.youtubeUrl = youtubeUrl;
-    if (boardAddress !== data.fetchBoard.boardAddress) {
+      formData.boardAddress?.zipcode !== data?.fetchBoard.boardAddress?.zipcode ||
+      formData.boardAddress?.address !== data?.fetchBoard.boardAddress?.address ||
+      formData.boardAddress?.addressDetail !== data?.fetchBoard.boardAddress?.addressDetail
+    ) {
       updateBoardInput.boardAddress = {
-        zipcode: zipcode,
-        address: address,
-        addressDetail: addressDetail,
+        zipcode: formData.boardAddress?.zipcode,
+        address: formData.boardAddress?.address,
+        addressDetail: formData.boardAddress?.addressDetail,
       };
     }
-    if (images.filter(Boolean)) {
-      updateBoardInput.images = images.filter(Boolean) as string[];
-      console.log(updateBoardInput);
+    const newImages = formData.images?.filter(Boolean);
+    if (JSON.stringify(newImages) !== JSON.stringify(data?.fetchBoard.images)) {
+      updateBoardInput.images = newImages as string[];
     }
-    console.log(updateBoardInput);
-    console.log(params.boardId);
 
     try {
-      const password = prompt(
-        "글을 입력할때 입력하셨던 비밀번호를 입력해주세요"
-      );
-      console.log(password);
-
+      const password = prompt("글을 입력할 때 입력하셨던 비밀번호를 입력해주세요");
       const result = await updateBoard({
-        variables: {
-          updateBoardInput,
-          password,
-          boardId: params.boardId as string,
-        },
+        variables: { updateBoardInput, password, boardId },
         refetchQueries: [
           {
             query: FetchBoardDocument,
-            variables: { boardId: params.boardId },
+            variables: { boardId },
           },
         ],
       });
-      console.log(result);
       router.push(`/boards/${result.data?.updateBoard._id}`);
     } catch (error) {
       const err = error as GraphQLError;
-      const showErrorModal = () =>
-        Modal.error({
-          title: "에러가 발생하였습니다.",
-          content: (err.message as string) ?? "에러가 발생하였습니다",
-        });
-      showErrorModal();
+      Modal.error({
+        title: "에러가 발생하였습니다.",
+        content: (err.message as string) ?? "에러가 발생하였습니다",
+      });
     }
   };
 
-  // 6. 취소 버튼 클릭 시 뒤로 가기
-  const onClickCancel = () => {
-    router.back();
-  };
-
   return {
-    inputs,
-    onChangeInputs,
-    onChangeBoardAddress,
-    onChangeYoutubeUrl,
-    onChangeFile,
-    onClickDelete,
-    onClickCancel,
-    isValid,
-    onClickUpdate,
+    register,
+    handleSubmit,
+    formState,
+    watch,
     onClickSubmit,
-    images,
     isModalOpen,
     onToggleModal,
+    onClickDelete,
+    onChangeFile,
+    onClickCancel,
     handleComplete,
-    setZipcode,
-    setAddress,
-    setAddressDetail,
-    boardAddress,
-    youtubeUrl,
+    onClickUpdate,
   };
 }
