--- maincamp/frontend/src/components/ui/button/Button.module.css
@@ -1,64 +1,0 @@
-.FormBtn,
-.CommentBtn {
-  width: fit-content;
-  height: 3rem;
-  padding: 0.75rem 1rem;
-  border-radius: 0.5rem;
-  border: none;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  gap: 0.5rem;
-
-  font-size: var(--font-size-btn);
-  font-weight: var(--font-weight-semibold);
-  text-align: center;
-  white-space: nowrap;
-}
-
-.FormBtn[type="button"],
-.CommentBtn[type="button"] {
-  border: 0.0625rem solid var(--color-gray-black);
-  background-color: var(--color-gray-white);
-
-  color: var(--color-gray-black);
-}
-
-.FormBtn[type="submit"],
-.modalBtn[type="submit"] {
-  background-color: var(--color-blue);
-
-  color: var(--color-gray-white);
-}
-
-.FormBtn[type="submit"]:disabled,
-.CommentBtn[type="submit"]:disabled {
-  background-color: var(--color-gray-300);
-  color: var(--color-gray-100);
-}
-
-.CommentBtn[type="submit"] {
-  background-color: var(--color-gray-black);
-  color: var(--color-gray-white);
-}
-
-.modalBtn {
-  height: 2.5rem;
-  min-width: 6rem;
-  max-width: 7.5rem;
-  padding: 0.5rem 0.75rem;
-  border-radius: 0.5rem;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 0.5rem;
-  flex: 1 0 0;
-
-  font-size: 0.875rem;
-  font-style: normal;
-  font-weight: 600;
-  line-height: 1.25rem;
-}
