--- maincamp/frontend/src/components/ui/input/InputBoardAddress.module.css
@@ -1,22 +1,0 @@
-.postForm__address__group {
-  display: flex;
-  flex-direction: column;
-}
-
-.zipCode__group {
-  width: 13.75rem;
-  display: flex;
-  flex-direction: row;
-  align-items: flex-end;
-  gap: 0.5rem;
-}
-
-.zipCode__group input {
-  margin: 0;
-  flex: 1 0 0;
-}
-
-.postForm__address__group input:read-only {
-  background-color: transparent;
-  border: 0.06rem solid var(--color-gray-200);
-}
