--- maincamp/frontend/src/components/ui/input/InputImage.module.css
@@ -1,55 +1,0 @@
-.add__image label{
-    width: 10rem;
-    height: 10rem;
-    border-radius:  0.5rem;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 0.5rem;
-
-    background-color: var(--color-gray-50);
-
-    color: var(--color-gray-600);
-}
-
-.add__image img{
-    fill: var(--color-gray-600);
-    width: 1.35rem;
-    height: 1.35rem;
-}
-
-.add__image span{
-    color: var(--color-gray-600);
-}
-
-.textDesktop{
-    display: inline;
-}
-
-.textMobile{
-    display: none;
-}
-
-@media (max-width: 780px){
-    .add__image label{
-        max-width: 10rem;
-        min-width: 6.25rem;
-        width: 100%;
-        flex: 1 1 0; 
-    }
-    .textDesktop{
-        display: none;
-    }
-    .textMobile{
-        display: inline;
-    }
-}
-
-
-
-
-
-
-
