--- maincamp/frontend/src/components/ui/input/Inputfield.module.css
@@ -1,70 +1,0 @@
-.Formfield label,
-.Formfield span {
-  font-size: var(--font-size-md);
-  font-weight: var(--font-weight-medium);
-}
-
-.Formfield span {
-  margin-left: 0.25rem;
-  color: var(--color-red);
-}
-
-.Formfield input,
-.Formfield textarea,
-.Commentfield textarea {
-  max-width: 80rem;
-  width: 100%;
-  height: 3rem;
-  border-radius: 0.5rem;
-  border: 0.06rem solid var(--color-gray-200);
-  outline: none;
-  padding: 0.75rem 1rem;
-  margin-top: 0.5rem;
-  vertical-align: top;
-
-  font-size: var(--font-size-md);
-  font-weight: var(--font-weight-regular);
-}
-
-.Formfield textarea {
-  height: 21rem;
-  resize: none;
-}
-
-.Formfield input::placeholder,
-.Formfield textarea::placeholder,
-.Commentfield input::placeholder {
-  color: var(--color-gray-400);
-}
-
-.Formfield input:disabled {
-  background-color: var(#f2f2f2);
-  border: none;
-}
-
-.Commentfield textarea {
-  height: 9rem;
-  resize: none;
-  margin: 0 0;
-}
-
-.Formfield > div {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-}
-
-.Formfield input:read-only {
-  background-color: var(--color-gray-50);
-  border: none;
-}
-
-.Commentfield textarea:read-only {
-  background-color: var(--color-gray-50);
-  border: none;
-}
-
-.Formfield > span {
-  display: block;
-  margin-top: 0.5rem;
-}
